Explicitly specify tasks for default rake build
We just want it to run the following tasks for now:

  * Run any pending databse migrations
  * Run the specs (unit tests)
  * Run the main cucumber features (acceptance/regression test)
  * Run the work-in-progress cucumber features

The "cucumber:wip" task runs features and scenarios tagged as @wip
separately, and does not fail the build if they fail. In fact, it *does*
fail the build if they succeed, as a reminder to remove the @wip tag.