Add an empty controller
Once again, this fixes the previous error, but causes another:

  The action 'index' could not be found for ItemsController ...

Note how we're going in small steps, making just enough of a change each
time to fix the specific thing that the test complains about, rather
than trying to do everything at once. This is what Bryan Liles calls
"make it pass or change the message".