Extract the item checks to a matcher
This allows us to keep the knowledge about how an item is displayed in
one place, rather than duplicating it in any steps that need to check
(even with this one step definition there was duplication, with the same
check for both items).

If we later change the way items are shown (maybe by using a table
rather than an unordered list), we only have to make one change in the
tests, rather than having to find everywhere we look for items.