Hard-code part of the expected output
This is just a quick sanity check to make sure we're testing for the
right thing. Now we've confirmed that the test is working (it now finds
the first item, and fails looking for the second), we can go on to
implement it properly using the actual values from the database.