Start a failing spec for ItemsController#index
Start by just calling the action and checking the outcome we want.
Obviously this fails because we haven't set up anything up yet:

  undefined local variable or method `items' ...

Normally you wouldn't want to commit a failing spec, but I've done it
here to show the small steps we're taking. As a compromise, the failing
changes are in a branch that will be merged in once it's green, so at
least if we follow the git history down the first parent everything will
build cleanly.