Implement step to check item has been added
The "show_todo_item" matcher expects an item object, so we build (but
don't save) one to pass in. Rather than duplicate the name from the
earlier step, we extract it to an instance variable.

Obviously the scenario now fails because we haven't written the code to
add the irem yet, but we can see that it fails for the correct reason.