Show done items appropriately in view
This makes all the tests pass, but we aren't quite done. Looking at the
site in the browser we realise that although we've added the "done"
class, we haven't written any CSS rules yet to display done items
differently. This is a good reminder to actually look at the app now and
again, rather than just relying on tests!