Replace logic in view with two separate partials
Logic in views is generally a bad thing, as views can be hard to test in
isolation. We're going to replace the if-else logic in the item partial
with two separate partials, one for items which are done and another for
those which aren't.

Obviously this doesn't work yet, as there's no way for Rails to know
which partial to render. We'll fix that next.