Add failing spec for Item#to_partial_path
Rails uses the 'to_display_path' method on model instances in
collections to decide which partial should be used to display them. In a
more complicated model we might decide to use decorator classes to wrap
the model instances, to separate business logic from presentation, but
for simplicity we'll implement the method directly on the model.

The method simply needs to return the right partial name depending on
the value of 'done', to match the pair of partials we created earlier.